requirements.txt
patricia-trie